refactor(react-doctor): decompõe ComparisonPanel (resíduo pós-#338) (#356)
Fase 4 do épico #239 (rastreada por #339): zera o no-giant-component:104
do ComparisonPanel (431 linhas, corpo de ~327) com decomposição real,
comportamento preservado:

- DivergenceActionsPanel: ramo divergente inteiro — botões Ambíguo/Pular,
  CustomAnswerInput (com key doc|campo e a lógica de currentValue),
  bloco de veredito anterior (leva junto formatVerdictDisplay, usado só
  ali) e a linha comentário/Anotar/Sugerir. O estado suggestOpen e o
  SuggestFieldDialog ficam no painel; o filho recebe onSuggest.

- UnansweredNotice: cálculo de unanswered extraído para a função pura
  listUnansweredHumans em escopo de módulo (filtros da issue #247 ponto 3
  + dedup por respondent_id ?? id) + o JSX do aviso, com retorno null
  quando todos preencheram.

Nenhuma mudança nas props públicas do ComparisonPanel; comentários
explicativos movidos junto com o código.

Verificação: react-doctor 0.5.8 — ComparisonPanel some do report
(Maintainability 12→11 warnings; resta o no-giant-component pré-existente
do ComparePageRoute em page.tsx, fora do checklist da #356); typecheck
sem erros novos (18 pré-existentes nos fixtures de ComparePage.test);
vitest 863/863.
