+++ frontend/src/components/compare/ComparisonPanel.tsx
@@ -4,32 +4,16 @@ import { useEffect, useMemo, useRef, useState } from "react";
 import { ProgressDots } from "../coding/ProgressDots";
 import { AgreementGroup, type FieldEquivalencePair } from "./AgreementGroup";
 import { MultiOptionReview } from "./MultiOptionReview";
-import { CustomAnswerInput } from "./CustomAnswerInput";
+import { DivergenceActionsPanel } from "./DivergenceActionsPanel";
+import { UnansweredNotice } from "./UnansweredNotice";
 import { KeyboardHints } from "./KeyboardHints";
 import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
 import { Badge } from "@/components/ui/badge";
-import { cn, normalizeForComparison } from "@/lib/utils";
+import { normalizeForComparison } from "@/lib/utils";
 import { buildResponseGroupKeys } from "@/lib/equivalence";
 import { ArrowRight, CheckCircle2, MessageSquare, Lightbulb } from "lucide-react";
-import { AddNoteButton } from "@/components/shared/AddNoteButton";
 import { SuggestFieldDialog } from "@/components/stats/SuggestFieldDialog";
 import type { PydanticField } from "@/lib/types";
-
-function formatVerdictDisplay(verdict: string): string {
-  if (verdict.startsWith("{")) {
-    try {
-      const parsed = JSON.parse(verdict) as Record<string, boolean>;
-      const selected = Object.entries(parsed)
-        .filter(([, v]) => v)
-        .map(([k]) => k);
-      return selected.length > 0 ? selected.join(", ") : "(nenhuma)";
-    } catch {
-      // fallback
-    }
-  }
-  return verdict;
-}
 
 interface ComparisonResponse {
   id: string;
@@ -159,40 +143,6 @@ export function ComparisonPanel({
 
   const feedbackBadge = commentCount + suggestionCount;
 
-  // "Não preencheu este campo" (issue #247, ponto 3): respostas sem valor para
-  // o campo atual ficam fora dos cards (que só mostram quem respondeu), o que
-  // fazia parecer que "só o robô respondeu". Listamos quem deixou o campo em
-  // branco para o revisor entender a ausência. Três filtros:
-  //  - `answer === undefined`: o campo está vazio (complementar ao `!== undefined`
-  //    que os cards/stats usam para contar quem respondeu);
-  //  - `!isFieldStale`: só contam respondentes cujo schema TINHA este campo e
-  //    ainda assim não o preencheram — não respondentes de um schema antigo onde
-  //    o campo nem existia (ruído de versão, não uma omissão real);
-  //  - `respondent_type === "humano"`: a issue é sobre humanos sumindo da tela; um
-  //    LLM pode omitir um campo legitimamente (ex.: condicional não satisfeita),
-  //    então "Robô não preencheu" seria ruído.
-  // Deduplicamos por `respondent_id ?? id` — mesma chave que a contagem da página
-  // (page.tsx) — para que um respondente com duas respostas qualificadas em branco
-  // (dados legados / re-codificação) conte e apareça uma vez só, sem fundir
-  // anônimos distintos (respondent_id null cai no id, que é único por linha).
-  const unanswered = useMemo(() => {
-    const seen = new Set<string>();
-    const out: { name: string }[] = [];
-    for (const r of responses) {
-      if (
-        r.answer !== undefined ||
-        r.isFieldStale ||
-        r.respondent_type !== "humano"
-      )
-        continue;
-      const key = r.respondent_id ?? r.id;
-      if (seen.has(key)) continue;
-      seen.add(key);
-      out.push({ name: r.respondent_name });
-    }
-    return out;
-  }, [responses]);
-
   return (
     <div className="flex h-full flex-col">
       <div className="shrink-0 border-b px-4 py-1.5">
@@ -265,112 +215,22 @@ export function ComparisonPanel({
           />
         )}
 
-        {unanswered.length > 0 && (
-          <div className="mt-2 rounded-md border border-dashed border-muted-foreground/20 bg-muted/30 px-2.5 py-1.5 text-[11px] leading-tight text-muted-foreground">
-            {unanswered.length === 1
-              ? "1 respondente não preencheu este campo"
-              : `${unanswered.length} respondentes não preencheram este campo`}
-            : {unanswered.map((r) => r.name).join(", ")}
-          </div>
-        )}
+        <UnansweredNotice responses={responses} />
 
         {isDivergent ? (
-          <>
-            {!isMulti && (
-              <div className="mt-2 flex flex-wrap gap-1">
-                <Button
-                  variant="outline"
-                  size="sm"
-                  className={cn(
-                    existingVerdict?.verdict === "ambiguo" &&
-                      "border-brand bg-brand/10 text-brand",
-                  )}
-                  onClick={() => onVerdict("ambiguo")}
-                >
-                  [A] Ambiguo
-                </Button>
-                <Button
-                  variant="outline"
-                  size="sm"
-                  className={cn(
-                    existingVerdict?.verdict === "pular" &&
-                      "border-brand bg-brand/10 text-brand",
-                  )}
-                  onClick={() => onVerdict("pular")}
-                >
-                  [S] Pular
-                </Button>
-                {/*
-                  "Nenhuma correta" + input de resposta nova (issue #247, ponto
-                  4). Keyed por doc|campo: navegar remonta e reseta o estado
-                  interno (aberto/valor) sem reset-em-effect — react-doctor só
-                  aceita key={identidade} para reset-on-prop-change.
-
-                  currentValue: no bloco !isMulti, um veredito de texto sem
-                  chosenResponseId que não seja marcador especial é, por
-                  construção, uma resposta custom (voto sempre carrega
-                  chosenResponseId). Passá-lo destaca o botão e re-semeia o
-                  input ao revisitar o campo — paridade com Ambíguo/Pular.
-                */}
-                <CustomAnswerInput
-                  key={`${documentId}|${fieldName}`}
-                  currentValue={
-                    existingVerdict &&
-                    existingVerdict.verdict !== "ambiguo" &&
-                    existingVerdict.verdict !== "pular" &&
-                    !existingVerdict.chosenResponseId
-                      ? existingVerdict.verdict
-                      : null
-                  }
-                  onSubmit={(value) => onVerdict(value)}
-                />
-              </div>
-            )}
-
-            {existingVerdict && (
-              <div className="mt-2 rounded-md bg-muted/50 px-3 py-1.5 text-xs text-muted-foreground">
-                Veredito anterior:{" "}
-                <span className="font-medium text-foreground">
-                  {formatVerdictDisplay(existingVerdict.verdict)}
-                </span>
-                {existingVerdict.comment && (
-                  <span className="ml-1">
-                    &mdash; &ldquo;{existingVerdict.comment}&rdquo;
-                  </span>
-                )}
-              </div>
-            )}
-
-            <div className="mt-2 flex flex-wrap items-center gap-2">
-              <Input
-                placeholder="Comentário (opcional)"
-                value={comment}
-                onChange={(e) => onCommentChange(e.target.value)}
-                className="flex-1 min-w-[180px] text-sm"
-              />
-              <AddNoteButton
-                key={documentId}
-                projectId={projectId}
-                documentId={documentId}
-                documentTitle={documentTitle}
-                fieldName={fieldName}
-                fieldLabel={fieldDescription}
-                variant="outline"
-                size="sm"
-                label="Anotar"
-              />
-              <Button
-                variant="outline"
-                size="sm"
-                className="gap-1"
-                onClick={() => setSuggestOpen(true)}
-                title="Sugerir alteração ao codebook neste campo"
-              >
-                <Lightbulb className="size-3.5" />
-                Sugerir
-              </Button>
-            </div>
-          </>
+          <DivergenceActionsPanel
+            projectId={projectId}
+            documentId={documentId}
+            documentTitle={documentTitle}
+            fieldName={fieldName}
+            fieldDescription={fieldDescription}
+            isMulti={!!isMulti}
+            existingVerdict={existingVerdict}
+            onVerdict={onVerdict}
+            comment={comment}
+            onCommentChange={onCommentChange}
+            onSuggest={() => setSuggestOpen(true)}
+          />
         ) : (
           <div className="mt-2 flex items-center justify-between gap-2 rounded-md border border-green-500/20 bg-green-500/5 px-3 py-2 text-xs text-muted-foreground">
             <div className="flex items-center gap-1.5">
